Refactored state to use new reducer
Couldnt get validation to work in reducer so this probably isn't the intended  way

Either way it was a fun example to implement a reducer and I think it cleaned up the code a tiny bit

Was able to do overall var name and code cleanup alongside above
