+++ src/components/UserInput/formReducer.jsx
@@ -1,10 +1,6 @@
-import { DateTime, Settings } from "luxon";
-Settings.throwOnInvalid = true;
-
 export const ACTION = {
-  Update: "update",
+  UpdateForm: "update-form",
   UpdateField: "update-field",
-  Validate: "validate",
   ClearFormError: "clear-error",
 };
 
@@ -31,7 +27,7 @@ export const initialState = [
 
 export function formReducer(formState, action) {
   switch (action.type) {
-    case ACTION.Update:
+    case ACTION.UpdateForm:
       return action.payload.formState;
     case ACTION.UpdateField:
       const predicate = (field) => field.name === action.payload.field.name;
@@ -47,83 +43,7 @@ export function formReducer(formState, action) {
           errorMsg: null,
         };
       });
-    case ACTION.Validate:
-      // Check individual fields
-      const validatedFields = validateFields(formState);
-
-      if (validatedFields.find((field) => field.hasError === true)) {
-        return validatedFields;
-      }
-
-      // Check if user provided values make a valid date
-      let userDate;
-      try {
-        userDate = DateTime.fromObject({
-          year: validatedFields.find((field) => field.name === "year").value,
-          month: validatedFields.find((field) => field.name === "month").value,
-          day: validatedFields.find((field) => field.name === "day").value,
-        });
-      } catch (error) {
-        const predicate = (field) => field.name === "day";
-        return validatedFields.toSpliced(validatedFields.findIndex(predicate), 1, {
-          ...validatedFields.find(predicate),
-          hasError: true,
-          errorMsg: "Must be a valid date",
-        });
-      }
-
-      // Check if user provided date is in the future
-      if (userDate > DateTime.now()) {
-        const predicate = (field) => field.name === "year";
-        return validatedFields.toSpliced(validatedFields.findIndex(predicate), 1, {
-          ...validatedFields.find(predicate),
-          hasError: true,
-          errorMsg: "Must be in the past",
-        });
-      }
-
-      // Otherwise return the validated fields to be saved into state
-      return validatedFields;
     default:
       return formState;
   }
-}
-
-function validateFields(fields) {
-  return fields.map((field) => {
-    const { name, value } = field;
-    if (value === "" || value === null) {
-      return {
-        ...field,
-        hasError: true,
-        errorMsg: "This field is required",
-      };
-    }
-
-    if (name === "day" && (value < 1 || value > 31)) {
-      return {
-        ...field,
-        hasError: true,
-        errorMsg: "Must be a valid day",
-      };
-    }
-
-    if (name === "month" && (value < 1 || value > 12)) {
-      return {
-        ...field,
-        hasError: true,
-        errorMsg: "Must be a valid month",
-      };
-    }
-
-    if (name === "year" && value < 1900) {
-      return {
-        ...field,
-        hasError: true,
-        errorMsg: "Year must be after 1900",
-      };
-    }
-
-    return field;
-  });
 }
